Fixed bug with display
Added functions to the OS to abstract display away from the
Processes. This way Processes do not need to know how to display
and output to the screen and can ask the operating system to do
it for them. Likewise the OS does not know how to display to the
screen, it just knows the module that does.