Changed directory methods in File_IO and OS
The directory traversal was being done in a iterative way. This
method may have worked, but is not robust. Directory traversal
should be handled recursively. This cleaned up some of
the code and made file searching more robust.

Testing of this new method has been done, but one implication is
that usage of OS.FS methods requires the use of either the
absolute path, or the relative path: e.g. "routes.csv" must now
be refered to as "/routes" or "./routes.csv"